refactor(test) move fixtures to separate file for re-use
