+++ spec/plugins/aws-lambda/99-access_spec.lua
@@ -2,100 +2,12 @@ local cjson   = require "cjson"
 local helpers = require "spec.helpers"
 local meta    = require "kong.meta"
 local pl_file = require "pl.file"
-
+local fixtures = require "spec.plugins.aws-lambda.fixtures"
 
 local TEST_CONF = helpers.test_conf
 local server_tokens = meta._SERVER_TOKENS
 local null = ngx.null
 
-
-local fixtures = {
-  dns_mock = helpers.dns_mock.new(),
-  http_mock = {
-    lambda_plugin = [[
-
-      server {
-          server_name mock_aws_lambda;
-          listen 10001 ssl;
-> if ssl_cert[1] then
-> for i = 1, #ssl_cert do
-          ssl_certificate     $(ssl_cert[i]);
-          ssl_certificate_key $(ssl_cert_key[i]);
-> end
-> else
-          ssl_certificate ${{SSL_CERT}};
-          ssl_certificate_key ${{SSL_CERT_KEY}};
-> end
-          ssl_protocols TLSv1.1 TLSv1.2 TLSv1.3;
-
-          location ~ "/2015-03-31/functions/(?:[^/])*/invocations" {
-              content_by_lua_block {
-                local function x()
-                  local function say(res, status)
-                    ngx.header["x-amzn-RequestId"] = "foo"
-
-                    if string.match(ngx.var.uri, "functionWithUnhandledError") then
-                      ngx.header["X-Amz-Function-Error"] = "Unhandled"
-                    end
-
-                    ngx.status = status
-
-                    if string.match(ngx.var.uri, "functionWithBadJSON") then
-                      local badRes = "{\"foo\":\"bar\""
-                      ngx.header["Content-Length"] = #badRes + 1
-                      ngx.say(badRes)
-
-                    elseif string.match(ngx.var.uri, "functionWithNoResponse") then
-                      ngx.header["Content-Length"] = 0
-
-                    elseif string.match(ngx.var.uri, "functionWithBase64EncodedResponse") then
-                      ngx.say("{\"statusCode\": 200, \"body\": \"dGVzdA==\", \"isBase64Encoded\": true}")
-
-                    elseif type(res) == 'string' then
-                      ngx.header["Content-Length"] = #res + 1
-                      ngx.say(res)
-
-                    else
-                      ngx.req.discard_body()
-                      ngx.header['Content-Length'] = 0
-                    end
-
-                    ngx.exit(0)
-                  end
-
-                  ngx.sleep(.2) -- mock some network latency
-
-                  local invocation_type = ngx.var.http_x_amz_invocation_type
-                  if invocation_type == 'Event' then
-                    say(nil, 202)
-
-                  elseif invocation_type == 'DryRun' then
-                    say(nil, 204)
-                  end
-
-                  local qargs = ngx.req.get_uri_args()
-                  ngx.req.read_body()
-                  local args = require("cjson").decode(ngx.req.get_body_data())
-
-                  say(ngx.req.get_body_data(), 200)
-                end
-                local ok, err = pcall(x)
-                if not ok then
-                  ngx.log(ngx.ERR, "Mock error: ", err)
-                end
-              }
-          }
-      }
-
-    ]]
-  },
-}
-
-
-fixtures.dns_mock:A {
-  name = "lambda.us-east-1.amazonaws.com",
-  address = "127.0.0.1",
-}
 
 
 for _, strategy in helpers.each_strategy() do
